Moved locking functions outside of the watch service.
